refactor(mocks): extract agentSource module from mockApi (#991 M3)
Move agentSource/agentSourceRemote state, the semver/status helpers (parseAgentSourceSemver, compareAgentSourceSemver, deriveAgentSourceStatus, agentSourceView), the mockAllowedAgentSourceHosts allowlist, and the five agent-source methods (getAgentSource, syncAgentSource, applyAgentSource, resolveAgentSourceLatest, updateCheckAgentSource) into agentSource.ts.

Per D6 the release-check helpers/methods stay in index.ts with settings (they read appSettings). agentSource and mockAllowedAgentSourceHosts are exported and imported back because updateSettings (still in index.ts) reads them.
